Refactoring Log
This is the first step of a refactoring of logging.
I added the name of the component that caused the entry to the log.
Unit tests now also check if the right component is handed over.

In the second step, the message class will be adapted such that messages don't contain the name of the component.