Add annotation for setting own enum values
Until now, the possible enum values are taken from the names of the components of a constant transformed to camel case.
Now, we want to allow setting own enum values by using a new annoation.

In a first step, I extended the ABAP Doc Parser so that the new annotation `$enumValue 'value' ` can be used.